fix test and rename files
