--- mlflow/store/_unity_catalog/oss_registry/__init__.py
@@ -1,1 +1,0 @@
-from mlflow.store._unity_catalog.oss_registry import oss_rest_store  # noqa: F401
