+++ mlflow/store/_unity_catalog/registry_oss/__init__.py
@@ -1,0 +1,1 @@
+from mlflow.store._unity_catalog.registry_oss import rest_store_oss  # noqa: F401
