+++ tests/utils/test_unity_catalog_oss_utils.py
@@ -4,7 +4,7 @@ from mlflow.protos.unity_catalog_oss_messages_pb2 import (
     RegisteredModelInfo,
     TagKeyValue,
 )
-from mlflow.utils._oss_unity_catalog_utils import registered_model_from_uc_oss_proto
+from mlflow.utils._unity_catalog_oss_utils import registered_model_from_uc_oss_proto
 
 
 def test_registered_model_from_uc_oss_proto():
